hotkeys are opt-in to minimize conflicts
This is a (somewhat) breaking change. Existing users now need to
explicitly provide the `-r` flag to enable hotkeys in interactive
terminals.

Motive: the use of terminal raw mode creates conflicts when running
in non-interactive environments such as containers, or when running
multiple `gow` concurrently. Making it opt-in avoids such conflicts
by default, and both the readme and the CLI help make it clear when
not to use it.